+++ src/Optimisers/HilbertianProjection.jl
@@ -202,6 +202,7 @@ struct HilbertianProjection{A} <: Optimiser
   - `ls_ξ_reduce_abs_tol = 0.01`: Tolerance on constraints to reduce `ls_ξ` via `ls_ξ_reduce_coef`.
   - `ls_γ_min = 0.001`: Minimum coeffient on the time step size for solving the HJ evolution equation.
   - `ls_γ_max = 0.1`: Maximum coeffient on the time step size for solving the HJ evolution equation.
+  - `ls_enable_it = 1`: Enable line search after this many iterations.
 
   A more concervative evolution of the boundary can be achieved by decreasing `ls_γ_max`.
 
@@ -222,7 +223,7 @@ struct HilbertianProjection{A} <: Optimiser
     λ=0.5, α_min=0.1, α_max=1.0, γ=0.1, reinit_mod = 1,
     ls_enabled = true, ls_max_iters = 10, ls_δ_inc = 1.1, ls_δ_dec = 0.7,
     ls_ξ = 1, ls_ξ_reduce_coef = 0.0025, ls_ξ_reduce_abs_tol = 0.01,
-    ls_γ_min = 0.001, ls_γ_max = 0.1,
+    ls_γ_min = 0.001, ls_γ_max = 0.1, ls_enable_it = 1,
     maxiter = 1000, verbose=false, constraint_names = map(i -> Symbol("C_$i"),1:N),
     converged::Function = default_hp_converged, debug = false,
     has_oscillations::Function = (ls_enabled ? (args...)->false : default_has_oscillations),
@@ -245,7 +246,7 @@ struct HilbertianProjection{A} <: Optimiser
       all(@. typeof(problem.analytic_dC) <: Function)) ? NoAutoDiff : WithAutoDiff;
 
     params = (;debug,γ,reinit_mod,ls_enabled,ls_max_iters,ls_δ_inc,ls_δ_dec,ls_ξ,
-               ls_ξ_reduce_coef,ls_ξ_reduce_abs_tol,ls_γ_min,ls_γ_max,os_γ_mult)
+               ls_ξ_reduce_coef,ls_ξ_reduce_abs_tol,ls_γ_min,ls_γ_max,os_γ_mult,ls_enable_it)
     new{A}(problem,ls_evolver,vel_ext,projector,history,converged,has_oscillations,params,φ0)
   end
 end
@@ -363,7 +364,7 @@ function _linesearch!(m::HilbertianProjection{WithAutoDiff},state,γ)
   it,J,C,θ,dJ,dC,uh,φh,uhd,vel,φ_tmp,_,os_it,reinit_cache,evo_cache = state
 
   params = m.params; history = m.history
-  ls_enabled = params.ls_enabled; reinit_mod = params.reinit_mod
+  ls_enabled, ls_enable_it = params.ls_enabled, params.ls_enable_it; reinit_mod = params.reinit_mod
   ls_max_iters,δ_inc,δ_dec = params.ls_max_iters,params.ls_δ_inc,params.ls_δ_dec
   ξ, ξ_reduce, ξ_reduce_tol = params.ls_ξ, params.ls_ξ_reduce_coef, params.ls_ξ_reduce_abs_tol
   γ_min, γ_max = params.ls_γ_min,params.ls_γ_max
@@ -377,7 +378,7 @@ function _linesearch!(m::HilbertianProjection{WithAutoDiff},state,γ)
       _,reinit_cache = reinit!(m.ls_evolver,φ,reinit_cache)
     end
 
-    ~ls_enabled && break
+    ~ls_enabled || (it <= ls_enable_it) && break
     ls_it += 1;
 
     # Calcuate new objective and constraints
@@ -408,7 +409,7 @@ function _linesearch!(m::HilbertianProjection{NoAutoDiff},state,γ)
   it,J,C,θ,dJ,dC,uh,φh,uhd,vel,φ_tmp,_,os_it,reinit_cache,evo_cache = state
 
   params = m.params; history = m.history
-  ls_enabled = params.ls_enabled; reinit_mod = params.reinit_mod
+  ls_enabled, ls_enable_it = params.ls_enabled, params.ls_enable_it; reinit_mod = params.reinit_mod
   ls_max_iters,δ_inc,δ_dec = params.ls_max_iters,params.ls_δ_inc,params.ls_δ_dec
   ξ, ξ_reduce, ξ_reduce_tol = params.ls_ξ, params.ls_ξ_reduce_coef, params.ls_ξ_reduce_abs_tol
   γ_min, γ_max = params.ls_γ_min,params.ls_γ_max
@@ -423,7 +424,7 @@ function _linesearch!(m::HilbertianProjection{NoAutoDiff},state,γ)
     end
 
     # Check enabled
-    if ~ls_enabled
+    if ~ls_enabled || (it <= ls_enable_it)
       J, C, dJ, dC = Gridap.evaluate!(m.problem,φh)
       return J, C, dJ, dC, γ, reinit_cache, evo_cache
     end